+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Atlas Content Modeler Changelog
 
 ## Unreleased
-
 ### Added
 - Related entries in relationship fields now have an “Edit” link to view or edit them in a new tab.
+
+### Fixed
+- Fields will now save empty values when existing content is removed.
 
 ## 0.19.0 - 2022-06-29
 ### Added